io_mvp_1/ui/requirements.txt
streamlit
pandas
numpy
openpyxl
xlsxwriter
python-dotenv
requests
plotly